### 2020-06-18 #2361 @whedon
[PRE REVIEW]: pvpumpingsystem: a python package for modeling and sizing photovoltaic water pumping systems
**Submitting author:** @tylunel (<a href="http://orcid.org/0000-0003-3008-1422">Tanguy Lunel</a>)
**Repository:** <a href="https://github.com/tylunel/pvpumpingsystem" target ="_blank">https://github.com/tylunel/pvpumpingsystem</a>
**Version:** v1.0
