### 2020-06-18 5f2388f
- standard_simulation.py changed to reflect the same pvps set-up than in test_sizing - general cleaning of sizing.py and test_sizing.py
