+++ pvpumpingsystem/sizing.py
@@ -11,14 +11,6 @@ import pandas as pd
 import tqdm
 import warnings
 
-import pvlib
-
-import pvpumpingsystem.pump as pp
-import pvpumpingsystem.pipenetwork as pn
-import pvpumpingsystem.consumption as cs
-import pvpumpingsystem.pvpumpsystem as pvps
-import pvpumpingsystem.mppt as mppt
-import pvpumpingsystem.reservoir as res
 import pvpumpingsystem.pvgeneration as pvgen
 
 
@@ -161,19 +153,19 @@ def subset_respecting_llp_direct(pv_database, pump_database,  # noqa: C901
         pvps_fixture.coupling = 'direct'
         warnings.warn("Pvps coupling method changed to 'direct'.")
 
-#    def funct_llp_for_Ms_Mp(pvps, M_s, M_p, **kwargs):
-#        pvps.pvgeneration.system.modules_per_string = M_s
-#        pvps.pvgeneration.system.strings_per_inverter = M_p
-#        pvps.pvgeneration.run_model()
-#        pvps.run_model(**kwargs)
-#        return pvps.llp
+    # def funct_llp_for_Ms_Mp(pvps, M_s, M_p, **kwargs):
+    #     pvps.pvgeneration.system.modules_per_string = M_s
+    #     pvps.pvgeneration.system.strings_per_inverter = M_p
+    #     pvps.pvgeneration.run_model()
+    #     pvps.run_model(**kwargs)
+    #     return pvps.llp
 
     # initialization of variables
     preselection = pd.DataFrame()
 
-#    pvps_fixture.pvgeneration.weather_data_and_metadata = {
-#            'weather_data': weather_data,
-#            'weather_metadata': weather_metadata}
+    # pvps_fixture.pvgeneration.weather_data_and_metadata = {
+    #         'weather_data': weather_data,
+    #         'weather_metadata': weather_metadata}
 
     for pv_mod_name in tqdm.tqdm(pv_database,
                                  desc='PV database exploration: ',
@@ -184,7 +176,32 @@ def subset_respecting_llp_direct(pv_database, pump_database,  # noqa: C901
             weather_data_and_metadata={
                     'weather_data': weather_data,
                     'weather_metadata': weather_metadata},
-            pv_module_name=pv_mod_name
+            pv_module_name=pv_mod_name,
+
+            price_per_watt=2.5,  # in US dollars
+            surface_tilt=45,  # 0 = horizontal, 90 = vertical
+            surface_azimuth=180,  # 180 = South, 90 = East
+            albedo=0,  # between 0 and 1
+            modules_per_string=1,
+            strings_in_parallel=1,
+            # PV module glazing parameters (not always given in specs)
+            glass_params={'K': 4,  # extinction coefficient [1/m]
+                          'L': 0.002,  # thickness [m]
+                          'n': 1.526},  # refractive index
+            racking_model='open_rack',  # or'close_mount' or 'insulated_back'
+
+            # Models used (check pvlib.modelchain for all available models)
+            orientation_strategy='south_at_latitude_tilt',  # or 'flat' or None
+            clearsky_model='ineichen',
+            transposition_model='isotropic',
+            solar_position_method='nrel_numpy',
+            airmass_model='kastenyoung1989',
+            dc_model='desoto',  # 'desoto' or 'cec'.
+            ac_model='pvwatts',
+            aoi_model='physical',
+            spectral_model='first_solar',
+            temperature_model='sapm',
+            losses_model='pvwatts'
             )
 
         for pump in tqdm.tqdm(pump_database,
@@ -229,60 +246,60 @@ def subset_respecting_llp_direct(pv_database, pump_database,  # noqa: C901
[…]
