### 2020-06-18 250d563
General cleaning of code
- standard_simulation.py rechanged to represent a more standard example (weather file changed)

- Commented old code of sizing and if '__name__'=='__main__': code removed
