+++ pvpumpingsystem/sizing.py
@@ -153,19 +153,8 @@ def subset_respecting_llp_direct(pv_database, pump_database,  # noqa: C901
         pvps_fixture.coupling = 'direct'
         warnings.warn("Pvps coupling method changed to 'direct'.")
 
-    # def funct_llp_for_Ms_Mp(pvps, M_s, M_p, **kwargs):
-    #     pvps.pvgeneration.system.modules_per_string = M_s
-    #     pvps.pvgeneration.system.strings_per_inverter = M_p
-    #     pvps.pvgeneration.run_model()
-    #     pvps.run_model(**kwargs)
-    #     return pvps.llp
-
     # initialization of variables
     preselection = pd.DataFrame()
-
-    # pvps_fixture.pvgeneration.weather_data_and_metadata = {
-    #         'weather_data': weather_data,
-    #         'weather_metadata': weather_metadata}
 
     for pv_mod_name in tqdm.tqdm(pv_database,
                                  desc='PV database exploration: ',
@@ -246,61 +235,6 @@ def subset_respecting_llp_direct(pv_database, pump_database,  # noqa: C901
                     pvps_fixture, llp_accepted,
                     M_s_min, M_s_max, M_p_min, M_p_max,)
 
-            # # Guess a M_s to start with or take the given M_s_guess:
-            # if M_s_guess is None:
-            #     M_s = M_s_min
-            # else:
-            #     M_s = M_s_guess
-            # # Guess a M_p to start with or take the given M_p_guess:
-            # if M_p_guess is None:
-            #     M_p = M_p_min
-            # else:
-            #     M_p = M_p_guess
-
-            # # initialization of variable for first round
-            # llp_prev_Ms = 1.1
-            # llp_prev_Mp = 1.1
-            # M_s_prev = M_s_min
-
-            # while True:
-            #     llp = funct_llp_for_Ms_Mp(pvps_fixture, M_s, M_p)
-            #     print('module: {0} / pump: {1} / M_s: {2} /'
-            #           '/ M_p: {3} / llp: {4} / npv: {5}'.format(
-            #             pv_mod_name, pump.idname, M_s, M_p,
-            #             llp, pvps_fixture.npv))
-            #     # changing M_s
-            #     if llp <= llp_accepted:
-            #         break
-            #     elif llp > llp_accepted:
-            #         if llp < llp_prev_Ms and M_s < M_s_max:
-            #             M_s_prev = M_s
-            #             M_s += 1
-            #             M_s_change_efficient = True
-            #         elif llp >= llp_prev_Ms:
-            #             M_s = M_s_prev
-            #             M_s_change_efficient = False
-            #         else:
-            #             pass  # keep current M_s
-            #     llp_prev_Ms = llp
-
-            #     if not M_s_change_efficient:  # then change M_p
-            #         llp = funct_llp_for_Ms_Mp(pvps_fixture, M_s, M_p)
-            #         print('module: {0} / pump: {1} / M_s: {2} /'
-            #               '/ M_p: {3} / llp: {4} / npv: {5}'.format(
-            #                       pv_mod_name, pump.idname, M_s, M_p,
-            #                       llp, pvps_fixture.npv))
-            #         # changing M_p
-            #         if llp <= llp_accepted:
-            #             break
-            #         elif llp > llp_accepted:
-            #             if llp < llp_prev_Mp and M_p < M_p_max:
-            #                 M_p += 1
-            #             elif M_p == M_p_max:
-            #                 break
-            #             else:
-            #                 break
-            #         llp_prev_Mp = llp
-
             preselection = preselection.append(
[…]
