### 2020-06-19 #2361 @kbarnhart
@danielskatz I'm overloaded at the moment, but expect to be able to take on 1-2 new joss submissions by the end of next week. Happy to take this one on at that point if it is still in need of an editor.
