### 2020-06-30 #2361 @kbarnhart
@tylunel thank you for your submission to JOSS. I've placed the paused label on this because there are a couple of issues that need to be addressed before the submission can move forward in the JOSS review process. I've outlined these below with a checklist. If you have questions or clarifications about these issues, please comment here and I'll do my best to provide guidance.

- [ ] Installation instructions should include how to install pvpumpingsytem itself.
- [ ] Documentation should be expanded (see [Review Criteria: Documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation)). This part of the JOSS review criteria documentation states:

    > "There should be sufficient documentation for you, the reviewer to understand the core functionality of the software under review. A high-level overview of this documentation should be included in a README file (or equivalent). "

    You are encouraged to accomplish this however you think is best. A common approach is a ReadTheDocs page. Given your extensive docstrings, this is likely the easiest solution.

    I would recommend that this documentation include the following three elements:
    * a narrative presentation of the package (statement of the core functionality, types of problems to be addressed, and/or design of the package),
    * the package API, and
    * example usage.
- [ ] State how a user might test the software.
