### 2020-07-06 #2361 @kbarnhart
@tylunel I wanted to follow up and see if you had any questions or clarifications about my comments.
