### 2020-07-08 #2361 @tylunel
Hi @kbarnhart, thank you for your review!
Actually I did not expect the first review to be done so quickly, thanks for that! And the reason why I am not very responsive now is that I am currently on vacation until july 15 haha. So I will really take over the submission only then, sorry for the inconvenience.
Your comments seem clear so far, just maybe one question: for the third checkbox "State how a user might test the software", do you mean "how a user can run the tests of the software" or more "how the user can have a glimpse of what she/he can do with the software"?
Thank you very much!
