### 2020-07-08 #2361 @kbarnhart
@tylunel no problem. Enjoy your vacation and let me know if you have additional questions for me when you return to this.

By "State how a user might test the software", I mean your first option. To state how a user can run the tests of the software. In the case of your contribution, I can see that there are test, just no instructions on how a person would run them.

It also looks like your CI badge only goes to an image of the badge, rather than a record of the CI.
