### 2020-08-05 012a470
Modifications of setup.py file in order to add the dependency 'fluids'.
