+++ setup.py
@@ -14,5 +14,5 @@ setup(
    url='https://github.com/tylunel/pvpumpingsystem',
    packages=['pvpumpingsystem'],  # same as name
    # external dependencies packages
-   install_requires=['numpy', 'pandas', 'scipy', 'pvlib'],
+   install_requires=['numpy', 'pandas', 'scipy', 'pvlib', 'fluids'],
    )
