### 2020-08-05 f8b807f
Modifications of setup.py file in order to add the dependency 'numpy-financial'.
