### 2020-08-05 b903a54
Modifications of README.md to improve the installation explanations.
