### 2020-08-05 83cb8e2
Style improvement of README.md
