+++ README.md
@@ -1,7 +1,7 @@
 ![CI general checks](https://github.com/tylunel/pvpumpingsystem/workflows/CI%20general%20checks/badge.svg?branch=master)
 
 # pvpumpingsystem
-"pvpumpingsystem" is a package providing tools for modeling and sizing photovoltaic water pumping systems.
+*pvpumpingsystem* is a package providing tools for modeling and sizing photovoltaic water pumping systems.
 
 It allows to find the quantity of water pumped from the pumping station characteristic given,
 or to choose some elements of the pumping station depending on the consumption of water needed.
@@ -10,7 +10,7 @@ It relies on pvlib-python for the photovoltaic power generation, and implements
 different models of pump for the simulation. 
 
 # Installation 
-pvpumpingsystem works with Python 3.5 and superior only.
+*pvpumpingsystem* works with Python 3.5 and superior only.
 
 ## For active users
 
@@ -19,18 +19,18 @@ If you are new to Python we recommend to use Anaconda to install and use Python.
 If you plan on editing the software and you are on Windows, install Git to simplify the download and the versioning of the package (on Linux, Git should be native).
 
 Once you have Anaconda and git installed, open the command line interface 'Anaconda prompt', change directory to the one you want to install pvpumpingsystem in, and type in:
-'''
+```
 pip install -e git+https://github.com/tylunel/pvpumpingsystem#egg=pvpumpingsystem
-'''
+```
 
 
 To ensure 'pvpumpingsystem' and its dependencies are properly installed, run the tests by going to the directory of pvpumpingsystem and by running pytest:
-'''
+```
 cd src/pvpumpingsystem
 pytest
-'''
+```
 
-## For passive users
+## For more passive users
 ----Still to come----
 
 # Hands-on start
