+++ docs/source/index.rst
@@ -1,0 +1,49 @@
+.. pvpumpingsystem documentation master file, created by sphinx-quickstart
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+
+Welcome to pvpumpingsystem's documentation!
+===========================================
+
+*pvpumpingsystem* is a package providing tools for modeling and sizing 
+photovoltaic water pumping systems.
+
+It allows to find the quantity of water pumped from the pumping station 
+characteristics given, or to choose some elements of the pumping station 
+depending on the consumption of water needed. It relies on pvlib-python 
+for the photovoltaic power generation, and implements different models 
+of pump for the simulation.
+
+The source code for pvpumpingsystem is hosted on GitHub:
+<https://github.com/tylunel/pvpumpingsystem>`_
+
+The package was originally developped at T3E research group, in Ecole de 
+Technologie Superieure, Montreal, Qc, Canada, by Tanguy Lunel.
+
+The software is published under the open source license GPL-v3.
+
+
+Citing pvpumpingsystem
+======================
+-----still to come------ 
+Paper currently under review at Journal of Open Source Software.
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+   
+   package_overview
+   installation
+   introtutorial
+   api
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
