+++ docs/source/package_overview.rst
@@ -1,0 +1,48 @@
+.. _package_overview:
+
+Package Overview
+================
+
+Introduction
+------------
+
+Pvpumpingsystem is an open source package providing various tools aimed 
+at facilitating the modeling and the sizing of photovoltaic powered water 
+pumping systems. Even though the package is originally targeted at researchers 
+and engineers, two practical examples are provided in order to help anyone 
+to use pvpumpingsystem.
+
+Python is the programming language used in the software, and the code is 
+structured within an object-oriented approach. Continuous integration 
+services allow checking for lint in the code and to automatize the tests. 
+Each class and function are documented with reference to the 
+literature when applicable. 
+
+Pvpumpingsystem is released under a GPL-v3 license.
+
+Pvpumpingsystem relies on already existing packages for photovoltaic 
+and fluid mechanics modeling, namely *pvlib-python* and *fluids*. 
+*pvpumpingsystem*'s originality lies in the implementation of various 
+motor-pump models for finite power sources and in the coupling 
+of the distinct component models.
+
+
+
+Getting support and contribute
+------------------------------
+
+If you need help, you think you have discovered a bug, or if you would
+like to edit *pvpumpingsystem*, then do not hesitate to open an issue on our 
+`GitHub issues page <https://github.com/tylunel/pvpumpingsystem/issues>`_
+or on our
+`GitHub pull request page <https://github.com/tylunel/pvpumpingsystem/pulls>`_ .
+
+
+Credits
+-------
+The T3E research group would like to acknowledge Mr. Michel Trottier for 
+his generous support, as well as the NSERC and the FRQNT for their grants 
+and subsidies. We also acknowledges the contributions and fruitful discussions 
+with Louis Lamarche and Sergio Gualteros that inspired and helped with the 
+current work.
+
