### 2020-08-06 1d104c0
Fix of conf.py file for documentation
