+++ docs/source/conf.py
@@ -32,11 +32,8 @@ release = '1.0'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-	'sphinx.ext.autodoc',
-    'sphinx.ext.autosummary']
-
-# enable autosummary option
-#autosummary_generate = True
+  'sphinx.ext.autodoc',
+  'sphinx.ext.autosummary']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
