### 2020-08-06 de052e0
Fix of conf.py file for documentation
