### 2020-08-06 81ce239
Fix of conf.py file for documentation
