### 2020-08-07 #2361 @tylunel
@kbarnhart I revised my package and added documentation with readthedocs.
Actually I still plan to expand a bit the _installation_ part next week (I must tell that I am not yet a real IT guy so I must dig a bit deeper into the subject to be able to give clear instructions to install pvpumpingsystem and its dependencies on different OS). So far the instruction are relevant only if the users have Anaconda and Git like me.

When you wrote that the docs should include examples, do you mean like jupyter notebook files? So far, I had put 2 normal python files (.py) in the folder `pvpumpingsystem/examples` with extensive docstring, is it enough?
