### 2020-08-07 ba91171
- Modification of README, and addition of a badge for the docs build. - Fix lints in conf.py
