+++ README.md
@@ -1,4 +1,5 @@
-![CI general checks](https://github.com/tylunel/pvpumpingsystem/workflows/CI%20general%20checks/badge.svg?branch=master)
+![CI general checks](https://github.com/tylunel/pvpumpingsystem/workflows/CI%20general%20checks/badge.svg)
+[![Documentation Status](https://readthedocs.org/projects/pvpumpingsystem/badge/?version=latest)](https://pvpumpingsystem.readthedocs.io/en/latest/?badge=latest)
 
 # pvpumpingsystem
 *pvpumpingsystem* is a package providing tools for modeling and sizing photovoltaic water pumping systems.
@@ -9,29 +10,27 @@ or to choose some elements of the pumping station depending on the consumption o
 It relies on pvlib-python for the photovoltaic power generation, and implements
 different models of pump for the simulation. 
 
+
 # Installation 
 *pvpumpingsystem* works with Python 3.5 and superior only.
 
-## For active users
+## With Anaconda and Git already installed
 
-If you are new to Python we recommend to use Anaconda to install and use Python. You can find it here: https://www.anaconda.com/products/individual.
-
-If you plan on editing the software and you are on Windows, install Git to simplify the download and the versioning of the package (on Linux, Git should be native).
-
-Once you have Anaconda and git installed, open the command line interface 'Anaconda prompt', change directory to the one you want to install pvpumpingsystem in, and type in:
+In *Anaconda prompt*, change directory to the one you want to install pvpumpingsystem in, and type in:
 ```
 pip install -e git+https://github.com/tylunel/pvpumpingsystem#egg=pvpumpingsystem
 ```
 
 
-To ensure 'pvpumpingsystem' and its dependencies are properly installed, run the tests by going to the directory of pvpumpingsystem and by running pytest:
+To ensure *pvpumpingsystem* and its dependencies are properly installed, run the tests by going to the directory of pvpumpingsystem and by running pytest:
 ```
 cd src/pvpumpingsystem
 pytest
 ```
 
-## For more passive users
-----Still to come----
+## Other installation
+Consult the docs for more information: https://pvpumpingsystem.readthedocs.io/en/latest/installation.html
+
 
 # Hands-on start
 
