### 2020-08-07 f4079a6
- build status badge now links to the CI page and not the badge image
