### 2020-08-07 0250aa3
Documentation expanded:
- 'introtutorial' renamed in 'beginnerguide'

- 'beginnerguide' expanded and detailed

- minor typo fix in 'index.rst'
