### 2020-08-07 c14ebfe
Fix of documentation links in index.rst
