+++ docs/source/index.rst
@@ -32,8 +32,8 @@ The software is published under the open source license GPL-v3.
    
    package_overview
    installation
-   introtutorial
-   api
+   beginnerguide
+   apireference
 
 
 
