### 2020-08-10 ac16faf
Update of dependencies in setup.py file (for installation without anaconda)
