+++ docs/source/environment.yml
@@ -2,12 +2,38 @@ name: pvpumpingsystemdocs
 
 dependencies:
     - python>=3.5
-    - numpy
-    - scipy
-    - pandas
     - matplotlib
-    - sphinx=3.0.4
     - pvlib>=0.7
     - numpy-financial
     - fluids
+    - tqdm
+    - sklearn
+    - sphinx=3.0.4
 
+depndencies of the dependencies:
+    - numpy
+    - scipy
+    - pandas
+    - requests
+    - python-dateutil
+    - certifi
+    - urllib3
+    - chardet
+    - idna
+    - pytz
+    - joblib
+    - threadpoolctl
+    - scikit-learn
+    - pillow
+    - kiwisolver
+    - cycler
+    - toml
+    - pluggy
+    - more-itertools
+    - py
+    - colorama
+    - attrs
+    - pyparsing
+    - packaging
+    - atomicwrites
+    - iniconfig 
