### 2020-08-10 0fb7675
Documentation expanded:
- installation.rst expanded with instruction for simple installation without anaconda or git.

- examples folder moved to docs
