### 2020-08-10 #2361 @kbarnhart
@tylunel thanks for making these improvements to pvpumpingsystem. However, there is still a bit more work that needs to be done before the package can move on to the review process.

Specifically, I think that the documentation and examples still need improvement.
- _Narrative documentation_: I think the Package Overview section you've written is a wonderful addition to the documentation, but would benefit from one or two more sentences describing the scope of problems pvpumpingsystem is intended to solve. A prospective user should be able to quickly use this information to determine whether they want to use pvpumpingsystem in their application.
- _API documentation_: The readthedocs page presently lists classes, functions, and methods provided by the package, but no information is provided regarding the inputs, outputs, and/or methods of each function and/or class. I see that you have documented much of this information in the source code docstrings, but I would expect to see this information compiled in the documentation pages. In addition, there are a number of instances in which an input is a path (e.g., for Pump, or for PVGeneration), but there is no documentation about the expected file format. The API level documentation should describe to a user the required inputs and outputs of all (best practice) or core (good practice) functions, classes, and methods.
- _Examples_: JOSS has no requirements regarding the specific form of examples (e.g., python scripts vs. notebooks) in order to support developers in creating examples they think work best for their code. However, I would expect that the examples would provide more background of the problem being solved and description of what the code is doing. When I look at the two example scripts, it takes me quite a while to figure out what exactly the problem being solved is. Adding text around the code (as comments) that introduces the problem and explains how pvpumpingsytem can be used to solve it is necessary. I would also recommend making it easy for a user to see the examples (e.g., direct link from a read the docs page to the python script, notebook viewer, or binder).

If you have questions or clarifications about these issues, please comment here and I'll do my best to provide guidance.
