### 2020-08-10 3bf42b5
Tests of sizing.py expanded
