+++ .github/workflows/pythonapp.yml
@@ -22,7 +22,7 @@ jobs:
 
     - name: Install dependencies
       run: |       # The pipe (|) is a tool allowing us to write multiple lines that execute independently without having to use multiple run commands or multiple jobs.
-        python -m pip install --upgrade pip
+        python -m pip install --upgrade pip  # update pip before using it
         # pip install git+https://github.com/tylunel/pvlib-python@all_changes#egg=pvlib  # Useful for testing with unreleased changes on pvlib-python. Specifies which branch to download with @<branch>
         pip install git+https://github.com/tylunel/pvpumpingsystem#egg=pvpumpingsystem
         pip install pvlib matplotlib sklearn tqdm fluids numpy_financial
