### 2020-08-11 d7087c7
Minor changes:
- Fix the version number in setup.py

- Docstring modification to MPPT()

- order of modules import changed to make it more understandable
