### 2020-08-11 83abab5
Documentation fixed and expanded
- 'beginnerguide' page readability improved with attributes marked as code and other minor changes

- 'standard_simulation' derived in a Jupyter Notebook with improved explanation
