+++ docs/source/beginnerguide.rst
@@ -3,9 +3,9 @@
 Beginner Guide
 ==============
 
-This page aims at explaining the global functioning of pvpumpingsystem. 
-Two examples are provided in `\pvpumpingsystem\pvpumpingsystem\examples` 
-are also available with extensive docstring in order to quickly understand 
+This page aims at explaining the global functioning of pvpumpingsystem.
+Two examples are provided in `\pvpumpingsystem\pvpumpingsystem\examples`
+are also available with extensive docstring in order to quickly understand
 core features of *pvpumpingsystem*.
 
 
@@ -15,18 +15,21 @@ Modeling paradigms
 ------------------
 
 *pvpumpingsystem* strucure is based on object-oriented programming.
-For users unfamiliar with the object-oriented programming, we recommend to invest
-some time to understand and discover the power of this programming paradigm.
+For users unfamiliar with the object-oriented programming, we recommend
+to invest some time to understand and discover the power of this programming
+paradigm.
 
-In *pvpumpingsystem*, in order to increase the understandability of the code, the physical components 
-of the PV pumping system corresponds to a class when possible, like for example 
-the classes Pump(), MPPT(), PipeNetwork(), Reservoir() and PVGeneration(). 
-Moreover, each of these classes are gathered into separate modules with 
+In *pvpumpingsystem*, in order to increase the understandability of the code,
+the physical components of the PV pumping system corresponds to a class
+when possible, like for example the classes Pump(), MPPT(), PipeNetwork(),
+Reservoir() and PVGeneration().
+Moreover, each of these classes are gathered into separate modules with
 appropriate names (`pump.py`, `mppt.py`, etc).
-The previous objects are then gathered in the class PVPumpSystem() which allows 
-running partial or comprehensive modeling of the pumping system. 
+The previous objects are then gathered in the class PVPumpSystem() which
+allows running partial or comprehensive modeling of the pumping system.
 
-A separate module `sizing.py` is dedicated to methods allowing to size these systems.
+A separate module `sizing.py` is dedicated to methods allowing to size these
+systems.
 
 
 
@@ -38,87 +41,114 @@ Simulation
 Inputs
 ^^^^^^
 
-In order to model a system, the simulation must take as input (minimum input to inform are
-in bold, the rest have default values):
+In order to model a system, the simulation must take as input:
+
 
 - PV generation characterictics:
-This part is basically a wrapper of pvlib-python functionnalities. The only new feature added
-is the automatic selection of the PV module reference which is the closest to the string
-given in `pv_module_name` attribute.
 
-  - `**pv_module_name**`: reference of the PV module
-  - `**weather_data_and_metadata**`: weather file. Can be any format supported by pvlib-python
-  - `surface_tilt`, `surface_azimuth`, `price_per_watt`,`modules_per_string`,`strings_in_parallel`, 
-  `glass_params`, and other options like the models to use can be given as well.
+This part is basically a wrapper of pvlib-python functionnalities. The only
+new feature added is the automatic selection of the PV module reference which
+is the closest to the string given in ``pv_module_name`` attribute.
+
+  - ``pv_module_name``: reference of the PV module
+
+  - ``weather_data_and_metadata**``: weather file. Can be any format supported
+    by pvlib-python
+
+  - optional: ``surface_tilt``, ``surface_azimuth``, ``price_per_watt``,
+    ``modules_per_string``, ``strings_in_parallel``, ``glass_params``,
+    and other options like the models to use can be given as well.
+
 
 - Electronics:
-In version 1.0, only DC motor-pump are considered, therefore only MPPT-DC/DC converters
-are considered (no inverter). It is called mppt, and it is actually used only if the coupling
-method chosen is 'mppt'.
 
-  - `efficiency`, `price` are the 2 main attributes.
[…]
