### 2020-08-12 f17a853
Documentation improved
- Package overview expanded with details on scope and application of the software
