### 2020-08-13 798b655
Docstrings improvement and minor fix
- Docstring of PVGeneration details better the file format accepted.

- Docstring of Pipenetwork completed

- Minor fix in pvpumpingsystem in case where no consumption is given
