### 2020-08-13 4cb6b01
Examples expanded: basic and easy example added
- Renaming of 'standard_simulation' into 'simulation_tunis_advanced'

- file 'simulation_tunis_basic.py' and 'simulation_tunis_basic.ipynb' created.

- '.ipynb_checkpoints' file removed from .gitignore
