### 2020-08-13 b85d1b3
Documentation of examples expanded:
- minor changes in code of 'simulation_tunis_basic'

- documentation of corresponding jupyter notebook file expanded
