### 2020-08-13 21d3377
Advanced example of simulation expanded
- 'simulation_tunis_advanced.py' expanded

- minor docstring improvement on consumption.py

- default value of the attribute 'spectral_model' changed to 'no_loss' in PVGeneration()
