+++ docs/examples/simulation_tunis_advanced.py
@@ -1,12 +1,15 @@
 # -*- coding: utf-8 -*-
 """
-Example of a simulation with pvpumpingsystem package.
+Example of a more advanced simulation with pvpumpingsystem package.
+The code is explained in details in the Jupyter Notebook file of the same
+name in the same folder.
 
 @author: Tanguy Lunel
 """
 
 import matplotlib.pyplot as plt
 import pandas as pd
+import pvlib
 
 import pvpumpingsystem.pvgeneration as pvgen
 import pvpumpingsystem.mppt as mppt
@@ -19,23 +22,39 @@ import pvpumpingsystem.pvpumpsystem as pvps
 # allows pandas to convert timestamp for matplotlib
 pd.plotting.register_matplotlib_converters()
 
-# ------------ LOCATION & PV MODELING ----------------------
+# ------------ LOCATION & WEATHER FILE IMPORT ---------------
+
+# Pvlib tools can be used to import the weather file wanted. All options can
+# be found at: https://pvlib-python.readthedocs.io/en/stable/api.html#io-tools
+
+# Import a weather file built by PVGIS according to latitude and longitude
+data, _, _, metadata = pvlib.iotools.get_pvgis_tmy(
+    lat=36,  # latitude of selected location
+    lon=10,  # longitude of selected location
+    outputformat='epw')  # output format of file
+
+# coerce the weather file to one single year to avoid bug with consumption file
+data.year = 2005
+data.index = data.index.map(lambda t: t.replace(year=2005))
+
+# ------------ PV MODELING ----------------------
 
 pvgen1 = pvgen.PVGeneration(
     # Weather data path
-    weather_data_and_metadata=(
-        '../../pvpumpingsystem/data/weather_files/'
-        'TUN_Tunis.607150_IWEC.epw'),  # to adapt:
+    weather_data_and_metadata={
+        'weather_data': data,
+        'weather_metadata': metadata},
 
     # PV array parameters
     pv_module_name='Canadian Solar CS5C 80M',
     price_per_watt=2.5,  # in US dollars,
     # The price must be given in dollar per watt
     # (if a 200W module costs 400USD, the price is 400/200 = 2 USD/W)
-    surface_tilt=45,  # 0 = horizontal, 90 = vertical
+    surface_tilt=50,  # 0 = horizontal, 90 = vertical
+    # More tilted to increase the yield in winter (at the expense of summer)
     surface_azimuth=180,  # 180 = South, 90 = East
     albedo=0.3,  # between 0 and 1
-    modules_per_string=4,
+    modules_per_string=5,  # 4 for mppt, 5 for direct (llp smaller, cheaper)
     strings_in_parallel=1,
     # PV module glazing parameters (not always given in specs)
     glass_params={'K': 4,  # extinction coefficient [1/m]
@@ -44,20 +63,23 @@ pvgen1 = pvgen.PVGeneration(
     racking_model='open_rack',  # or'close_mount' or 'insulated_back'
 
     # Models used (check pvlib.modelchain for all available models)
-    orientation_strategy='south_at_latitude_tilt',  # or 'flat' or None
+    orientation_strategy=None,  # or 'flat', or 'south_at_latitude_tilt'
     clearsky_model='ineichen',
     transposition_model='isotropic',
     solar_position_method='nrel_numpy',
     airmass_model='kastenyoung1989',
-    dc_model='desoto',  # 'desoto' or 'cec'.
+    dc_model='desoto',  # 'desoto' or 'cec', must be a single diode model.
     ac_model='pvwatts',
     aoi_model='physical',
-    spectral_model='first_solar',
+    spectral_model='no_loss',  # recommended when not sure if the file is
[…]
