### 2020-08-19 bfb78bc
Documentation expanded
- 'standard_sizing.py' docstrings improved
