+++ docs/examples/standard_sizing.py
@@ -16,15 +16,51 @@ import pvpumpingsystem.reservoir as rv
 import pvpumpingsystem.mppt as mppt
 from pvpumpingsystem import sizing
 
+# The sizing presented here is a sizing that chooses the best combination of
+# motor-pump and PV module to reduce the life cycle cost of a system.
+# It first computes the number of PV modules required for the combination
+# to pump enough water in order to respect the maximum water shortage
+# probability (named llp) accepted. Then it selects the combination with
+# the lowest net present value.
+# --------- ELEMENTS TO SIZE----------------------------------------------
 
-# ------------ MAIN INPUTS --------------------------------------------------
+# ------------- Pumps -----------
+# Three pumps are available here. The user wants to find the one which fits
+# best for the application. First the 3 pumps must be imported.
+pump_1 = pp.Pump(
+    path="../../pvpumpingsystem/data/pump_files/SCB_10_150_120_BL.txt")
+
+pump_2 = pp.Pump(
+    path="../../pvpumpingsystem/data/pump_files/SCB_10_150_180_BL.txt")
+
+# Note that this pump does not have complete information given in the file,
+# so some details are added here.
+pump_3 = pp.Pump(
+    path="../../pvpumpingsystem/data/pump_files/Shurflo_9325.txt",
+    idname='Shurflo_9325',
+    price=700,
+    motor_electrical_architecture='permanent_magnet')
+
+# The database must be provided under the form of a list for the sizing:
+pump_database = [pump_1,
+                 pump_2,
+                 pump_3]
+
+# ------------- PV Modules -----------
+# PV array database:
+pv_database = ['Kyocera solar KU270 6MCA',
+               'Canadian Solar CS5C 80M']
+
+
+# --------- MAIN INPUTS --------------------------------------------------
 
 # Weather input
-weather_path = (
-    '../data/weather_files/CAN_PQ_Montreal.Intl.AP.716270_CWEC.epw')
 weather_data, weather_metadata = pvlib.iotools.epw.read_epw(
-        weather_path, coerce_year=2005)
-# short weather to compute faster
+    '../../pvpumpingsystem/data/weather_files/TUN_Tunis.607150_IWEC.epw',
+    coerce_year=2005)
+
+# shorten the weather data by keeping the worst month only (based on GHI)
+# in order to compute faster.
 weather_data = sizing.shrink_weather_worst_month(weather_data)
 
 # Consumption input
@@ -33,14 +69,13 @@ consumption_data = cs.Consumption(constant_flow=5)  # in L/min
 # Pipes set-up
 pipes = pn.PipeNetwork(h_stat=20,  # vertical static head [m]
                        l_tot=100,  # length of pipes [m]
-                       diam=0.08,  # diameter of pipes [m]
-                       material='plastic',
-                       optimism=True)
+                       diam=0.05,  # diameter of pipes [m]
+                       material='plastic')
 
 # Reservoir
-reserv1 = rv.Reservoir(size=1000000,
-                       water_volume=0,
-                       price=1000)
+reservoir1 = rv.Reservoir(size=5000,  # [L]
+                          water_volume=0,  # [L] at beginning
+                          price=(1010+210))  # 210 is pipes price
 
 # MPPT
 mppt1 = mppt.MPPT(efficiency=0.96,
@@ -54,17 +89,11 @@ pvgen1 = pvgen.PVGeneration(
                     'weather_metadata': weather_metadata},  # to adapt:
 
             # PV array parameters
-            pv_module_name='kyocera solar KU270 6MCA',
+            pv_module_name=pv_database[0],
             price_per_watt=2.5,  # in US dollars
             surface_tilt=45,  # 0 = horizontal, 90 = vertical
[…]
