### 2020-08-20 04b0e9a
Documentation on API improved
- fix of API. API is now interactive, details can be provided on each function referenced in apireference.rst
