+++ docs/source/conf.py
@@ -49,6 +49,8 @@ master_doc = 'index'
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
+# Activate this to True in order to allow the autogeneration from docstrings
+autosummary_generate = True
 
 # -- Options for HTML output -------------------------------------------------
 
