+++ docs/source/generated/pvpumpingsystem.mppt.MPPT.rst
@@ -1,0 +1,22 @@
+pvpumpingsystem.mppt.MPPT
+=========================
+
+.. currentmodule:: pvpumpingsystem.mppt
+
+.. autoclass:: MPPT
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~MPPT.__init__
+   
+   
+
+   
+   
+   
