+++ docs/source/generated/pvpumpingsystem.pvpumpsystem.PVPumpSystem.calc_reservoir.rst
@@ -1,0 +1,6 @@
+pvpumpingsystem.pvpumpsystem.PVPumpSystem.calc\_reservoir
+=========================================================
+
+.. currentmodule:: pvpumpingsystem.pvpumpsystem
+
+.. automethod:: PVPumpSystem.calc_reservoir
