+++ docs/source/generated/pvpumpingsystem.pvpumpsystem.PVPumpSystem.calc_flow.rst
@@ -1,0 +1,6 @@
+pvpumpingsystem.pvpumpsystem.PVPumpSystem.calc\_flow
+====================================================
+
+.. currentmodule:: pvpumpingsystem.pvpumpsystem
+
+.. automethod:: PVPumpSystem.calc_flow
