+++ docs/source/generated/pvpumpingsystem.sizing.size_nb_pv_mppt.rst
@@ -1,0 +1,6 @@
+pvpumpingsystem.sizing.size\_nb\_pv\_mppt
+=========================================
+
+.. currentmodule:: pvpumpingsystem.sizing
+
+.. autofunction:: size_nb_pv_mppt
