+++ docs/source/generated/pvpumpingsystem.pump._curves_coeffs_theoretical_constant_efficiency.rst
@@ -1,0 +1,6 @@
+pvpumpingsystem.pump.\_curves\_coeffs\_theoretical\_constant\_efficiency
+========================================================================
+
+.. currentmodule:: pvpumpingsystem.pump
+
+.. autofunction:: _curves_coeffs_theoretical_constant_efficiency
