+++ docs/source/generated/pvpumpingsystem.finance.initial_investment.rst
@@ -1,0 +1,6 @@
+pvpumpingsystem.finance.initial\_investment
+===========================================
+
+.. currentmodule:: pvpumpingsystem.finance
+
+.. autofunction:: initial_investment
