+++ docs/source/generated/pvpumpingsystem.pump.Pump.iv_curve_data.rst
@@ -1,0 +1,6 @@
+pvpumpingsystem.pump.Pump.iv\_curve\_data
+=========================================
+
+.. currentmodule:: pvpumpingsystem.pump
+
+.. automethod:: Pump.iv_curve_data
