### 2020-08-20 6224bf9
Documentation expanded
- jupyter notebook created for 'sizing_example'

- 'standard_sizing' renamed in 'sizing_example'

- minor changes on 'simulation_tunis_advanced'
