### 2020-08-20 addace2
Improvement on structure of documentation
- 'beginnerguide.rst' removed, and the content transfered to package_overview or in the jupyter notebook examples
