### 2020-08-21 85bc4dd
Nbviewer and binder added
- nbviewer links to the examples added to README.md

- binder badge added to README.md

- Documentation modification in README and index.rst
