+++ docs/source/index.rst
@@ -6,21 +6,22 @@
 Welcome to pvpumpingsystem's documentation!
 ===========================================
 
-*pvpumpingsystem* is a package providing tools for modeling and sizing 
+*pvpumpingsystem* is a package providing tools for modeling and sizing
 offgrid photovoltaic water pumping systems. It is specially designed for
 small to medium size systems, the type of pumping system typically used
 for isolated communities.
 
-It allows to find the quantity of water pumped from the pumping station 
-characteristics given, or to choose some elements of the pumping station 
-depending on the consumption of water needed. It relies on pvlib-python 
-for the photovoltaic power generation, and implements different models 
-of pump for the simulation.
+It can model the whole functioning of such pumping system on an hourly basis
+and eventually provide key financial and technical findings on a year.
+Conversely it can help choose some elements of the pumping station
+depending on output values wanted (like daily water consumption and
+acceptable risk of water shortage). Find more on the scope of the software
+in the section :ref:`package_overview`.
 
 The source code for pvpumpingsystem is hosted on GitHub:
 https://github.com/tylunel/pvpumpingsystem
 
-The package was originally developped at T3E research group, in Ecole de 
+The package was originally developped at T3E research group, in Ecole de
 Technologie Superieure, Montreal, Qc, Canada, by Tanguy Lunel.
 
 The software is published under the open source license GPL-v3.
@@ -29,22 +30,14 @@ The software is published under the open source license GPL-v3.
 .. toctree::
    :maxdepth: 3
    :caption: Contents:
-   
+
    package_overview
    installation
    apireference
 
 
 
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
-
-
 Citing pvpumpingsystem
 ======================
------still to come------ 
+-----still to come------
 Paper currently under review at Journal of Open Source Software.
