### 2020-08-21 99cf06a
Documentation improved
- Named links added

- example.rst added to better reference the examples
