+++ docs/source/examples.rst
@@ -1,0 +1,35 @@
+.. _examples:
+
+Examples
+========
+
+Three examples of how the software can be used are in the folder
+``docs/examples``.
+
+
+Jupyter Notebook
+----------------
+
+For a given system, the first two show how to obtain the outflows,
+probability of water shortage, life cycle cost and many other results:
+
+`basic usage example <https://nbviewer.jupyter.org/github/tylunel/pvpumpingsystem/blob/master/docs/examples/simulation_tunis_basic.ipynb>`_
+
+`more advanced usage example <https://nbviewer.jupyter.org/github/tylunel/pvpumpingsystem/blob/master/docs/examples/simulation_tunis_advanced.ipynb>`_
+
+
+The third shows how to optimize the selection of one or more component
+on the pumping station based on user requirements.
+
+`sizing example <https://nbviewer.jupyter.org/github/tylunel/pvpumpingsystem/blob/master/docs/examples/sizing_example.ipynb>`_
+
+Previous examples can be run in binder as well by clicking on the
+corresponding icon in the upper right-hand corner of nbviewer pages, or by
+accessing through the
+`binder build <https://mybinder.org/v2/gh/tylunel/pvpumpingsystem/master>`_.
+
+
+Python files
+------------
+These examples are also available in the form of python files in order to
+freely adapt the code to your needs. Directly check out in ``docs/examples``.
