+++ docs/source/index.rst
@@ -33,6 +33,7 @@ The software is published under the open source license GPL-v3.
 
    package_overview
    installation
+   examples
    apireference
 
 
