### 2020-08-24 #2361 @kbarnhart
@tylunel thanks for making these changes. I definitely think that expanding your examples into notebooks and adding explanatory text substantially improves the package documentation. I also tried out the binder instance and it worked well! I find Binder is a great way to allow potential users to try out software quickly. Nice work 👍

Perhaps I'm mistaken, but I still don't see API documentation that describes the inputs and outputs of functions and classes.
I'm looking [here](https://pvpumpingsystem.readthedocs.io/en/latest/apireference.html#) in the documentation.

I see every function/class listed, but there is no information about what each does, what inputs are required and what outputs are provided. I want to make sure that I'm communicating my request clearly so I'll provide an example of what I'm looking for.

This [numpy function for the identity matrix](https://numpy.org/doc/stable/reference/generated/numpy.identity.html#numpy.identity
) has the following documentation:

It was built based on the docstring:

A common approach to making this sort of documentation is to use [sphinx autodoc](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html), formatted either based on the Numpy or Google convention (I've been pleased with using the [napoleon sphinx extension](https://sphinxcontrib-napoleon.readthedocs.io/en/latest/) for this). I also want to be clear that you don't need to use any specific method or set of tools to accomplish this (e.g., you don't need to use sphinx autodoc because I gave it as an example), but you do need to provide API level documentation for the package in some form.

If you have any questions or clarifications for me, please let me know and I'll do my best to provide guidance.
