### 2020-08-25 #2361 @tylunel
Ho sorry! Actually you were very clear on what you expected, and I did the changes accordingly. But it seems that the documentation builds correctly only locally, and not when I push it on GitHub and readthedocs. So I'm embarrassed, I forgot to check that my last changes on API build well online too... I will investigate why local and online builds differ on that.
