+++ docs/source/conf.py
@@ -33,7 +33,8 @@ release = '1.0'
 # ones.
 extensions = [
   'sphinx.ext.autodoc',
-  'sphinx.ext.autosummary']
+  'sphinx.ext.autosummary',
+  'sphinx.ext.napoleon']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
