### 2020-08-25 56f8787
Docstrings fixed to better respect the Numpy doc style
