### 2020-08-25 7398d8c
Attempt to fix API online: sphinx version changed in environment
