+++ docs/source/environment.yml
@@ -8,7 +8,7 @@ dependencies:
     - fluids
     - tqdm
     - sklearn
-    - sphinx=3.0.4
+    - sphinx==1.8.5
 
 depndencies of the dependencies:
     - numpy
