### 2020-08-26 8b18ddb
Attempt to fix API online:
- workaround with toctree listing all generated files
