+++ docs/source/api2.rst
@@ -1,0 +1,126 @@
+.. currentmodule:: pvpumpingsystem
+
+API reference 2
+===============
+
+
+Classes
+-------
+
+The different classes of *pvpumpingsystem*.
+
+.. toctree::
+
+   generated/pvpumpingsystem.pvgeneration.PVGeneration
+   generated/pvpumpingsystem.mppt.MPPT
+   generated/pvpumpingsystem.pump.Pump
+   generated/pvpumpingsystem.pipenetwork.PipeNetwork
+   generated/pvpumpingsystem.reservoir.Reservoir
+   generated/pvpumpingsystem.consumption.Consumption
+   generated/pvpumpingsystem.pvpumpsystem.PVPumpSystem
+
+
+Functions and methods
+---------------------
+
+Components modeling
+^^^^^^^^^^^^^^^^^^^
+
+.. toctree::
+
+   generated/pvpumpingsystem.reservoir.Reservoir.change_water_volume
+
+   generated/pvpumpingsystem.consumption.adapt_to_flow_pumped
+
+   generated/pvpumpingsystem.pipenetwork.PipeNetwork.dynamichead
+
+   generated/pvpumpingsystem.pvgeneration.PVGeneration.run_model
+
+   generated/pvpumpingsystem.pump.Pump.iv_curve_data
+   generated/pvpumpingsystem.pump.Pump.functIforVH
+   generated/pvpumpingsystem.pump.Pump.functIforVH_Arab
+   generated/pvpumpingsystem.pump.Pump.functIforVH_Kou
+   generated/pvpumpingsystem.pump.Pump.functIforVH_theoretical
+   generated/pvpumpingsystem.pump.Pump.functQforVH
+   generated/pvpumpingsystem.pump.Pump.functQforPH
+   generated/pvpumpingsystem.pump.Pump.functQforPH_Hamidat
+   generated/pvpumpingsystem.pump.Pump.functQforPH_Arab
+   generated/pvpumpingsystem.pump.Pump.functQforPH_Kou
+   generated/pvpumpingsystem.pump.Pump.functQforPH_theoretical
+   generated/pvpumpingsystem.pump.get_data_pump
+   generated/pvpumpingsystem.pump.specs_completeness
+   generated/pvpumpingsystem.pump._curves_coeffs_Arab06
+   generated/pvpumpingsystem.pump._curves_coeffs_Kou98
+   generated/pvpumpingsystem.pump._curves_coeffs_Hamidat08
+   generated/pvpumpingsystem.pump._curves_coeffs_theoretical
+   generated/pvpumpingsystem.pump._curves_coeffs_theoretical_variable_efficiency
+   generated/pvpumpingsystem.pump._curves_coeffs_theoretical_constant_efficiency
+   generated/pvpumpingsystem.pump._curves_coeffs_theoretical_basic
+   generated/pvpumpingsystem.pump._domain_V_H
+   generated/pvpumpingsystem.pump._domain_P_H
+   generated/pvpumpingsystem.pump._extrapolate_pow_eff_with_cst_efficiency
+   generated/pvpumpingsystem.pump.plot_Q_vs_P_H_3d
+   generated/pvpumpingsystem.pump.plot_I_vs_V_H_3d
+   generated/pvpumpingsystem.pump.plot_Q_vs_V_H_2d
+
+
+Global modeling
+^^^^^^^^^^^^^^^
+
+.. toctree::
+
+   generated/pvpumpingsystem.pvpumpsystem.PVPumpSystem.define_motorpump_model
+   generated/pvpumpingsystem.pvpumpsystem.PVPumpSystem.operating_point_noiteration
[…]
