+++ docs/source/conf.py
@@ -34,7 +34,15 @@ release = '1.0'
 extensions = [
   'sphinx.ext.autodoc',
   'sphinx.ext.autosummary',
-  'sphinx.ext.napoleon']
+  'sphinx.ext.napoleon',
+
+  'sphinx.ext.mathjax',
+  'sphinx.ext.viewcode',
+  'sphinx.ext.intersphinx',
+  'sphinx.ext.extlinks',
+  'IPython.sphinxext.ipython_directive',
+  'IPython.sphinxext.ipython_console_highlighting',
+  ]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
