### 2020-08-26 776fe9a
Attempt to fix API online:
- workaround v2, less extensions added
