+++ docs/source/conf.py
@@ -39,9 +39,7 @@ extensions = [
   'sphinx.ext.mathjax',
   'sphinx.ext.viewcode',
   'sphinx.ext.intersphinx',
-  'sphinx.ext.extlinks',
-  'IPython.sphinxext.ipython_directive',
-  'IPython.sphinxext.ipython_console_highlighting',
+  'sphinx.ext.extlinks'
   ]
 
 # Add any paths that contain templates here, relative to this directory.
