### 2020-08-26 7c2d2bb
Attempt to fix API online:
- workaround v2, sphinx 3 needed in conf file
