### 2020-08-26 f8b34f4
Attempt to fix API online:
- workaround v2 confirmed, and renaming of files
