+++ docs/source/conf.py
@@ -29,7 +29,7 @@ release = '1.0'
 # -- General configuration ---------------------------------------------------
 
 # If your documentation needs a minimal Sphinx version, state it here.
-needs_sphinx = '3.0'
+#needs_sphinx = '3.0'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -37,12 +37,7 @@ needs_sphinx = '3.0'
 extensions = [
   'sphinx.ext.autodoc',
   'sphinx.ext.autosummary',
-  'sphinx.ext.napoleon',
-
-  'sphinx.ext.mathjax',
-  'sphinx.ext.viewcode',
-  'sphinx.ext.intersphinx',
-  'sphinx.ext.extlinks'
+  'sphinx.ext.napoleon'
   ]
 
 # Add any paths that contain templates here, relative to this directory.
