### 2020-08-26 242a79c
Attempt to fix API online:
- autosummary with :toctree: modules
