+++ docs/source/api_autosummary.rst
@@ -4,6 +4,21 @@ API reference
 =============
 
 
+Modules
+-------
+
+.. autosummary::
+   :toctree: modules
+
+   pvgeneration
+   mppt
+   pump
+   pipenetwork
+   reservoir
+   consumption
+   pvpumpsystem
+
+
 Classes
 -------
 
