+++ docs/source/modules/pvpumpingsystem.pipenetwork.rst
@@ -1,0 +1,22 @@
+pvpumpingsystem.pipenetwork
+===========================
+
+.. automodule:: pvpumpingsystem.pipenetwork
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      PipeNetwork
+   
+   
+
+   
+   
+   
