+++ docs/source/modules/pvpumpingsystem.pvpumpsystem.rst
@@ -1,0 +1,33 @@
+pvpumpingsystem.pvpumpsystem
+============================
+
+.. automodule:: pvpumpingsystem.pvpumpsystem
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      calc_efficiency
+      calc_flow_directly_coupled
+      calc_flow_mppt_coupled
+      calc_reservoir
+      function_i_from_v
+      operating_point_noiteration
+   
+   
+
+   
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      PVPumpSystem
+   
+   
+
+   
+   
+   
