### 2020-08-31 #2361 @tylunel
@kbarnhart Now it should be okay!
As often in this kind of case, the problem was quite stupid but it still took me a long time to find out where it came from... Anyway, the API online is now as expected.
