### 2020-08-31 4d9e3b8
Fix API online:
- Settings of readthedocs account changed ('install pvpumpingsystem' ticked)

- minor clean of docstrings

- deletion of files built by former fix attempts
