+++ docs/source/modules/pvpumpingsystem.pump.rst
@@ -1,32 +1,0 @@
-pvpumpingsystem.pump
-====================
-
-.. automodule:: pvpumpingsystem.pump
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      get_data_pump
-      plot_I_vs_V_H_3d
-      plot_Q_vs_P_H_3d
-      plot_Q_vs_V_H_2d
-      specs_completeness
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      Pump
-   
-   
-
-   
-   
-   
