+++ docs/source/index.rst
@@ -34,7 +34,7 @@ The software is published under the open source license GPL-v3.
    package_overview
    installation
    examples
-   api
+   apireference
 
 
 
