+++ docs/source/modules/pvpumpingsystem.pvpumpsystem.PVPumpSystem.rst
@@ -1,28 +1,0 @@
-pvpumpingsystem.pvpumpsystem.PVPumpSystem
-=========================================
-
-.. currentmodule:: pvpumpingsystem.pvpumpsystem
-
-.. autoclass:: PVPumpSystem
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~PVPumpSystem.__init__
-      ~PVPumpSystem.calc_efficiency
-      ~PVPumpSystem.calc_flow
-      ~PVPumpSystem.calc_reservoir
-      ~PVPumpSystem.define_motorpump_model
-      ~PVPumpSystem.operating_point_noiteration
-      ~PVPumpSystem.run_model
-   
-   
-
-   
-   
-   
