+++ docs/source/modules/pvpumpingsystem.mppt.rst
@@ -1,22 +1,0 @@
-pvpumpingsystem.mppt
-====================
-
-.. automodule:: pvpumpingsystem.mppt
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      MPPT
-   
-   
-
-   
-   
-   
