+++ docs/source/modules/pvpumpingsystem.consumption.rst
@@ -1,28 +1,0 @@
-pvpumpingsystem.consumption
-===========================
-
-.. automodule:: pvpumpingsystem.consumption
-
-   
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      adapt_to_flow_pumped
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      Consumption
-   
-   
-
-   
-   
-   
