+++ docs/source/modules/pvpumpingsystem.reservoir.Reservoir.rst
@@ -1,23 +1,0 @@
-pvpumpingsystem.reservoir.Reservoir
-===================================
-
-.. currentmodule:: pvpumpingsystem.reservoir
-
-.. autoclass:: Reservoir
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~Reservoir.__init__
-      ~Reservoir.change_water_volume
-   
-   
-
-   
-   
-   
