+++ docs/source/modules/pvpumpingsystem.pipenetwork.rst
@@ -1,22 +1,0 @@
-pvpumpingsystem.pipenetwork
-===========================
-
-.. automodule:: pvpumpingsystem.pipenetwork
-
-   
-   
-   
-
-   
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      PipeNetwork
-   
-   
-
-   
-   
-   

[…]
