### 2020-09-01 #2361 @kbarnhart
Nice work @tylunel. Glad you were able to get the docs to build on readthedocs! They look great and I think this is a huge improvement already!

I've now un-paused this submission and will begin the stage of the JOSS review process in which I find two or three reviewers for this submission.

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @, so you would refer to me as @ kbarnhart instead of @kbarnhart). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

I will incorporate your recommendations into my identification of reviewers.
