### 2020-09-02 #2361 @tylunel
Thank you!
I would specially recommend @ motahhir who seems to already work on PV pumping systems. Then, from the 'topic/areas' column, @ kumar and @ robinroche could fit as well. Lastly, @ wholmgren may be also interested as _pvpumpingsystem_ uses his package _pvlib-python_, and that some of the structure/philosophy of my package was inspired by his.
