### 2020-09-03 #2637 @kbarnhart
@samuelduchesne and @robinroche thanks for agreeing to review this submission to JOSS. As noted above, we are trying to adjust to authors, reviewers, and editors needs under COVID. Thus we are asking reviewers try and complete reviews in 6 weeks. I am going to use whedon the editorial bot to remind you about your review in three weeks, half way through the requested timeline.

If there are any questions from the authors or reviewers throughout the process, please feel free to let me know here or directly at krbarnhart@usgs.gov.

Thanks again for contributing to the JOSS review process.
