### 2020-09-03 2736e63
Documentation improvement and minor change in financial parameters declaration:
- codecov badge added to README.md

- docstring improved in finance.py and pvpumpsystem.py

- all financial parameters must be declared in finance.py functions now (the 2 optional attributes of PVPumpSystem 'labour_price_coefficient' and 'discount_rate' removed)
