+++ pvpumpingsystem/pvpumpsystem.py
@@ -37,46 +37,41 @@ class PVPumpSystem(object):
 
     Attributes
     ----------
-        pvgeneration: pvpumpingsystem.PVGeneration,
-            //!\\ The weather file used here should not smooth the extreme
-            conditions (avoid TMY or IWEC).
-            //!\\ the pvgeneration.modelchain.dc_model must be a Single Diode
-            model if the system is directly-coupled
+    pvgeneration: pvpumpingsystem.PVGeneration,
+        Note that the weather file used here should ideally not
+        smooth the extreme conditions (avoid TMY or IWEC for example).
+        The pvgeneration.modelchain.dc_model must be a Single Diode
+        model if the system is directly-coupled
 
-        motorpump: pvpumpingsystem.Pump
-            The pump used in the system.
+    motorpump: pvpumpingsystem.Pump
+        The pump used in the system.
 
-        coupling: str,
-            represents the type of coupling between pv generator and pump.
-            Can be 'mppt' or 'direct'
+    coupling: str,
+        represents the type of coupling between pv generator and pump.
+        Can be 'mppt' or 'direct'
 
-        motorpump_model: str, default None
-            The modeling method used to model the motorpump. Can be:
-            'kou', 'arab', 'hamidat' or 'theoretical'.
-            Overwrite the motorpump.modeling_method attribute if not None.
+    motorpump_model: str, default None
+        The modeling method used to model the motorpump. Can be:
+        'kou', 'arab', 'hamidat' or 'theoretical'.
+        Overwrite the motorpump.modeling_method attribute if not None.
 
-        mppt: pvpumpingsystem.MPPT
-            Maximum power point tracker of the system.
+    mppt: pvpumpingsystem.MPPT
+        Maximum power point tracker of the system.
 
-        pipes: pvpumpingsystem.PipeNetwork
+    pipes: pvpumpingsystem.PipeNetwork
 
-        reservoir: pvpumpingsystem.Reservoir
+    reservoir: pvpumpingsystem.Reservoir
 
-        consumption: pvpumpingsystem.Consumption
+    consumption: pvpumpingsystem.Consumption
 
-        labour_price_coefficient: float, default is 0.20
-            Ratio of the price of labour and secondary costs (wires,
-            racks (can be expensive!), transport of materials, etc) on initial
-            investment. It is considered at 0.2 in Gualteros (2017),
-            but is more around 0.40 in Tarpuy(Peru) case.
+    llp: None or float,
+        Loss of Load Probability, i.e. Water shortage probability.
+        It is None until computed by run_model(), and then it
+        ranges between 0 and 1.
 
-        Computed attributes
-        -------------------
-            llp: float, between 0 and 1
-                Loss of Load Probability, i.e. Water shortage probability.
-
-            inital_investment: float
-                Cost of the system at the installation [USD]
+    initial_investment: None or float,
+        Cost of the system at the installation [USD].
+        It is None until computed by run_model()
 
     """
     def __init__(self,
@@ -88,15 +83,11 @@ class PVPumpSystem(object):
                  pipes=None,
                  reservoir=None,
                  consumption=None,
-                 labour_price_coefficient=0.20,  # TODO: realistic value?
-                 discount_rate=0.05,
                  idname=None):
         self.pvgeneration = pvgeneration  # instance of PVArray
         self.motorpump = motorpump  # instance of Pump
         self.coupling = coupling
         self.mppt = mppt
-        self.labour_price_coefficient = labour_price_coefficient
-        self.discount_rate = discount_rate
 
         if motorpump_model is None and motorpump is not None:
             self.motorpump_model = self.motorpump.modeling_method
@@ -145,7 +136,7 @@ class PVPumpSystem(object):
[…]
