+++ pvpumpingsystem/finance.py
@@ -15,7 +15,7 @@ import numpy as np
 import numpy_financial as npf
 
 
-def initial_investment(pvps, labour_price_coefficient=None):
+def initial_investment(pvps, labour_price_coefficient=0.2):
     """
     Function computing the initial investment cost.
 
@@ -24,23 +24,21 @@ def initial_investment(pvps, labour_price_coefficient=None):
     pvps: pvpumpingsystem.PVPumpSystem,
         The Photovoltaic pumping system whose cost is to be analyzed.
 
-    labour_price_coefficient: float, default is None
-            Ratio of the price of labour and secondary costs on initial
-            investment.
-            None value will look for value of pvps object. In pvps, default
-            is 1.2.
+    labour_price_coefficient: float, default is 0.2
+            Ratio of the price of labour and secondary costs (wires,
+            racks (can be expensive!), transport of materials, etc) on initial
+            investment. It is considered at 0.2 in Gualteros (2017),
+            but is more around 0.40 in Tarpuy(Peru) case.
 
     Returns
     -------
-    float: Initial investment for the whole pumping system.
+    float
+        Initial investment for the whole pumping system.
     """
     try:
         pv_modules_price = (pvps.pvgeneration.system.modules_per_string
                             * pvps.pvgeneration.system.strings_per_inverter
                             * pvps.pvgeneration.price_per_module)
-
-        if labour_price_coefficient is None:
-            labour_price_coefficient = pvps.labour_price_coefficient
 
         if pvps.coupling == 'mppt' or pvps.coupling == 'mppt_no_iteration':
             result = (pvps.motorpump.price
@@ -63,7 +61,8 @@ def initial_investment(pvps, labour_price_coefficient=None):
 
 
 # TODO: put lifespan of each object in the correponding object itself
-def net_present_value(pvps, opex, discount_rate=0.02,
+def net_present_value(pvps,  discount_rate=0.02,
+                      labour_price_coefficient=0.2, opex=0,
                       lifespan_pv=30, lifespan_mppt=10, lifespan_pump=10):
     """
     Function computing the net present value of a PVPS
@@ -73,25 +72,32 @@ def net_present_value(pvps, opex, discount_rate=0.02,
     pvps: pvpumpingsystem.PVPumpSystem,
         The photovoltaic pumping system to evaluate.
 
-    opex: float,
-        Yearly operational expenditure of the pvps.
-
     discount_rate: float, default is 0.02
         Dicsount rate.
 
-    lifespan_pv: float
-        Lifespan of the photovoltaic modules. It is also considered as the
-        lifespan of the whole system.
+    labour_price_coefficient: float, default is 0.2
+        Ratio of the price of labour on capital cost.
+            Example: If labour_price_coefficient = 0.2 (20%), it is
+            considered that a 1000 USD PV array will cost 200 USD more
+            to be installed on site.
 
-    lifespan_mppt: float
-        Lifespan of the mppt
+    opex: float,
+        Yearly operational expenditure of the pvps.
 
-    lifespan_pump: float
-        Lifespan of the pump
+    lifespan_pv: float, default is 30
+        Lifespan of the photovoltaic modules in years. It is also
+        considered as the lifespan of the whole system.
+
+    lifespan_mppt: float, default is 10
+        Lifespan of the mppt in years.
+
+    lifespan_pump: float, default is 10
+        Lifespan of the pump in years.
 
     Returns
     -------
-    float: The net present value of the PVPS
+    float:
+        The net present value of the PVPS
     """
     # creation of list of opex cost with length of lifespan_pv
     cashflow_list = [opex]
@@ -103,19 +109,19 @@ def net_present_value(pvps, opex, discount_rate=0.02,
                             * pvps.pvgeneration.price_per_module)
 
[…]
