### 2020-09-03 bdeb173
Documentation improvements
- Docstrings of Pump.py fixed to fit better the numpy style
