+++ pvpumpingsystem/pump.py
@@ -34,41 +34,39 @@ class Pump:
 
     Attributes
     ----------
-        path: str, default=''
-            The path to the .txt file with the pump specifications.
+    path: str, default=''
+        The path to the .txt file with the pump specifications.
 
-        motor_electrical_architecture: str, default is None
-            'permanent_magnet', 'series_excited', 'shunt_excited',
-            'separately_excited'.
+    motor_electrical_architecture: str, default is None
+        'permanent_magnet', 'series_excited', 'shunt_excited',
+        'separately_excited'.
 
-        modeling_method: str, , default is 'arab'
-            name of the method used for modeling the pump.
+    modeling_method: str, default is 'arab'
+        name of the method used for modeling the pump.
 
-        idname: str, default is None
-            name of the pump
+    idname: str, default is None
+        name of the pump
 
-        price: numeric, default is None
-            The price of the pump
+    price: numeric, default is None
+        The price of the pump
 
-        controller: str, default is None
-            Name of controller
+    controller: str, default is None
+        Name of controller
 
-    Computed Attributes: (available after the object declaration)
+    voltage_list: None or list,
+        list of voltage (the keys of preceding dictionaries) [V]
 
-        voltage_list: list,
-            list of voltage (the keys of preceding dictionaries) [V]
+    specs: None or pandas.DataFrame,
+        Dataframe with columns of following numeric:
+            'voltage': voltage at pump input [V]
+            'current': current at pump input [A]
+            'power': electrical power at pump input [W]]
+            'tdh': total dynamic head in the pipes at output [m]
+            'flow': pump output flow rate [liter per minute]
 
-        specs: pandas.DataFrame,
-            Dataframe with columns of following numeric:
-                'voltage': voltage at pump input [V]
-                'current': current at pump input [A]
-                'power': electrical power at pump input [W]]
-                'tdh': total dynamic head in the pipes at output [m]
-                'flow': pump output flow rate [liter per minute]
-
-        data_completeness: dict,
-            Provides some figures to assess the completeness of the data.
-            (for more details, see pump.specs_completeness() )
+    data_completeness: None or dict,
+        Provides some figures to assess the completeness of the data.
+        (for more details, see pump.specs_completeness() )
 
     """
     _ids = count(1)
@@ -232,9 +230,10 @@ class Pump:
 
         Return
         ------
-        dict with following couples keys:values :
-            I: list of current [A]
-            V: list of voltage [V]
+        dict
+            with following couples keys-values:
+                I: list of current [A]
+                V: list of voltage [V]
         """
 
         fctI, intervals = self.functIforVH()
@@ -256,13 +255,11 @@ class Pump:
 
         Returns
         -------
-        Tuple containing :
-
-        Function giving I according to voltage V and head H for the pump:
-            I = f1(V, H)
-
-        Domains of validity for V and H. Can be functions, so as the
-            range of one depends on the other, or fixed ranges.
+        tuple
+            - Function giving I according to voltage V and head H for the pump:
+                I = f1(V, H)
+            - Domains of validity for V and H. Can be functions, so as the
[…]
