### 2020-09-04 06a130d
Documentation improved
- All docstrings modified to better fit Numpy doc standards

- test added. Explicitly test the financial analysis of a pvps
