### 2020-09-07 0f82cda
Documentation improved and minor fix
- Jupyter notebook files completed to show how to change financial parameters

- Handling of financial parameters fixed in processes of sizing
