### 2020-09-25 5a6bb6e
Fix in examples provided in docs
- A multiplication per 60 had been forgotten in order to convert liters/min in liters/year

- Small explanation added to clarify the operating of pump when water tank full
