### 2020-09-25 #2637 @kbarnhart
@robinroche and @samuelduchesne thanks for checking in. As I noted above, the recent reminder marks three weeks into the review and JOSS is asking reviewers to complete reviews in six weeks. Thanks in advance for your contributions to the JOSS review process.
