### 2020-09-28 #2637 @robinroche
@tylunel @kbarnhart
I am done with my review.
I have some suggestions for improvement:

1. Installation
- "open the command line interface ‘Anaconda prompt’": briefly explain how, or link to a page explaining how.
- "cd src/pvpumpingsystem pytest": I am on linux and this command did not work as is. However, splitting it in two worked.
- More information on the automated test script should be provided to the user. What can the user expect to see when running it? I got 26 warnings (see below).
- "Besides the libraries contained in Anaconda, pvpumpingsystem also requires: * pvlib-python: * fluids: * numpy-financial:": there is a minor formatting issue with this list.
- I feel that a "Getting started" page would help right after installation, and describe the basic steps of the design of a simple system design. The examples are great, but I feel that this would make the new user more comfortable.

Documentation:
- Examples of target audiences are not listed in the documentation, unless I missed it.
- Explain briefly what each example does.
- What is the source of data on pumps, etc.?

Paper:
- There is no comparison with any comparable existing software. What is new/better in the proposed one?

---

Results of the test script:

============================================ test session starts ============================================
rootdir: /home/rroche/src/pvpumpingsystem
plugins: doctestplus-0.5.0, arraydiff-0.3, astropy-header-0.1.2, hypothesis-5.5.4, remotedata-0.3.2, openfiles-0.4.0
collected 16 items

pvpumpingsystem/test/test_pump.py .......                                                             [ 43%]
pvpumpingsystem/test/test_pvpumpsystem.py .....                                                       [ 75%]
pvpumpingsystem/test/test_sizing.py ....                                                              [100%]

============================================= warnings summary ==============================================
pvpumpingsystem/test/test_pump.py::test_init
pvpumpingsystem/test/test_pump.py::test_all_models_coeffs
pvpumpingsystem/test/test_pump.py::test_functIforVH
pvpumpingsystem/test/test_pump.py::test_functQforPH
pvpumpingsystem/test/test_pump.py::test_iv_curve_data
pvpumpingsystem/test/test_sizing.py::test_sizing_minimize_npv_mppt
pvpumpingsystem/test/test_sizing.py::test_sizing_minimize_npv_mppt
pvpumpingsystem/test/test_sizing.py::test_sizing_minimize_npv_direct
pvpumpingsystem/test/test_sizing.py::test_sizing_minimize_npv_direct
  /home/rroche/src/pvpumpingsystem/pvpumpingsystem/pump.py:102: UserWarning: idname attribute overwritten.
    warnings.warn('idname attribute overwritten.')

pvpumpingsystem/test/test_pump.py::test_init
pvpumpingsystem/test/test_pump.py::test_limited_pump_data_1
pvpumpingsystem/test/test_pump.py::test_all_models_coeffs
pvpumpingsystem/test/test_pump.py::test_functIforVH
pvpumpingsystem/test/test_pump.py::test_functQforPH
pvpumpingsystem/test/test_pump.py::test_iv_curve_data
  /home/rroche/src/pvpumpingsystem/pvpumpingsystem/pump.py:112: UserWarning: motor_electrical_architecture attribute overwritten.
    warnings.warn('motor_electrical_architecture '

pvpumpingsystem/test/test_pump.py::test_limited_pump_data_1
pvpumpingsystem/test/test_pump.py::test_limited_pump_data_2
  /home/rroche/src/pvpumpingsystem/pvpumpingsystem/pump.py:1515: UserWarning: Power and current data will be recomputedwith constant efficiency assumption.
    warnings.warn('Power and current data will be recomputed'

pvpumpingsystem/test/test_pump.py::test_limited_pump_data_2
  /home/rroche/src/pvpumpingsystem/pvpumpingsystem/pump.py:1297: UserWarning: Simplistic model of constant efficiency applied.Better model could be used if permanent_magnet
    warnings.warn('Simplistic model of constant efficiency applied.'

pvpumpingsystem/test/test_pump.py::test_functIforVH
  /home/rroche/src/pvpumpingsystem/pvpumpingsystem/inverse.py:199: RuntimeWarning: Trouble calculating inverse for values: [7.]
    "%s" % str(yin[~resultsmask]), RuntimeWarning)

pvpumpingsystem/test/test_pump.py::test_functIforVH
  /home/rroche/src/pvpumpingsystem/pvpumpingsystem/pump.py:381: RuntimeWarning: invalid value encountered in sqrt
    return R_a*i + beta*np.sqrt(i)

pvpumpingsystem/test/test_pump.py::test_functIforVH
  /home/rroche/src/pvpumpingsystem/pvpumpingsystem/inverse.py:207: RuntimeWarning: Results obtained with less than 2 decimal digits of accuracy
    % accuracy, RuntimeWarning)

pvpumpingsystem/test/test_sizing.py::test_sizing_minimize_npv_mppt
pvpumpingsystem/test/test_sizing.py::test_sizing_minimize_npv_direct
  /home/rroche/src/pvpumpingsystem/pvpumpingsystem/pump.py:94: UserWarning: price attribute overwritten.
    warnings.warn('price attribute overwritten.')

pvpumpingsystem/test/test_sizing.py::test_sizing_minimize_npv_mppt
pvpumpingsystem/test/test_sizing.py::test_sizing_minimize_npv_direct
  /home/rroche/anaconda3/lib/python3.7/site-packages/scipy/optimize/minpack.py:808: OptimizeWarning: Covariance of the parameters could not be estimated
    category=OptimizeWarning)

pvpumpingsystem/test/test_sizing.py::test_sizing_minimize_npv_direct
  /home/rroche/src/pvpumpingsystem/pvpumpingsystem/sizing.py:225: UserWarning: Pump Shurflo_9325 and PV module voltage do not match
    'do not match').format(pump.idname))

-- Docs: https://docs.pytest.org/en/latest/warnings.html
===================================== 16 passed, 26 warnings in 17.41s ======================================
