### 2020-09-28 #2637 @samuelduchesne
@kbarnhart, I am done with the review as well. On my end, I worked on MacOS and had no issues installing and using the tool from the installation documentation. I enjoyed trying out this package which is very well coded (excellent programming ethics). The examples were very useful to understand how to use the package. Having both jupyter notebooks and python scripts in the examples folder is a welcomed feature.

Regarding installation, my only suggestion would be to upload the package to PyPi (or to conda-forge) so that a regular `pip install` (or `conda install`) is possible; installing in editable mode is unusual for published packages.

I agree with @robinroche's suggestions for improvement which cover most of the package-related considerations. I do have general comments as well as some questions regarding the methodology the tool implements (which should not be overlooked...). I present them bellow.

### Scope

- [ ] The scope section (and the `readme`) needs a good diagram visually explaining the tool (the jupyter notebook example has a nice one). This will help users understand quickly the use of this tool.

### Mention of "minimization"
There is very little information in the docs as well as in the paper about the kind of optimization that is performed in the sizing module.

- [ ] How is the NPV minimized? It seems that only a parametric study is performed where some configurations are kept and others are discarded;
- [ ] Looking at the code, it seems that `sizing_minimize_npv()` returns a list of possible PV module and pump configurations. What if multiple systems satisfy the constraints? Then, the user must choose?

Also, the hardcoded values of 45° and 180° for the PV surface tilt and azimuth in `subset_respecting_llp_direct` is questionable; An optimal panel angle usually follows the latitude of the location.

- [ ] Why did the author chose these values?https://github.com/tylunel/pvpumpingsystem/blob/5a6bb6e1c3d1387e04a989cdf1a54baa956f7b50/pvpumpingsystem/sizing.py#L172-L173

- [ ] I suggest that the author more clearly define the two use cases in the documentation (and the paper): 1. the modeling of pv pumping system operation and 2. the sizing of such systems.

- [ ] Clearly mention the limitations of the sizing module. What are the variables that can (or cannot) be optimized?
