### 2020-09-29 #2637 @robinroche
@kbarnhart
Reading @samuelduchesne 's comments, I realize that I somehow forgot to paste my comments on the technical (application) and conclusions. However they are similar to @samuelduchesne 's. This a nice contribution that can be useful to both researchers and designers of such systems.
