### 2020-09-30 #2637 @tylunel
Thanks for you very constructive comments @samuelduchesne @robinroche ! I will address the points you raise and come back to you with the corresponding improvements.

But before I have two questions. The first is about the warnings in the tests:
> More information on the automated test script should be provided to the user. What can the user expect to see when running it? I got 26 warnings (see below)
I am not sure what is the best practice here, and unfortunately internet did not provide a clear answer on it either haha. I added several warnings in my code in order to guide the user during a normal simulation or sizing, but this results in having all these warnings in the tests as well. Should I try not to add as much `UserWarning` (in my view it would make the use of the software maybe more complicated for unexperienced user)? Or is it okay to ignore some class of warnings in the tests like `UserWarnings` (as implicitly suggested here: https://docs.pytest.org/en/stable/warnings.html)? Or is it enough just to tell the user that it is normal to see 26 warnings when she/he runs the tests?

The second question in actually more a request for clarification to @robinroche :
> I feel that a "Getting started" page would help right after installation, and describe the basic steps of the design of a simple system design. The examples are great, but I feel that this would make the new user more comfortable.
What does this "Getting started" page should contain more exactly in your opinion? Help on how to run the python and jupyter notebook examples? More general explanations around what is a PV pumping system (so far, I assume that the users have a minimum knowledge about PV pumping systems)? A more abstract layout of the steps of simulation than what is proposed in the examples? Thanks in advance for providing further details
