### 2020-09-30 #2637 @kbarnhart
@robinroche and @samuelduchesne Thanks for your thoughtful reviews! At this point in the JOSS process we will now let @tylunel address these issues. If more extensive discussion needs to take place I would recommend that @tylunel make an issue in the submitted repository, paste and link to the above review comments, and focus the discussion there. This makes it easier to keep track of multiple threads.

I'll weigh in briefly regarding two items.
1. I agree with @samuelduchesne that distributing via conda-forge or pip would be highly encouraged.
2. Regarding warnings and the result of tests I think the most important thing is to communicate to a user how to evaluate whether the tests passed or not. Secondly I think that if you know (as the developer) that a test will produce a warning (e.g., the example above where the warning is "UserWarning: price attribute overwritten.") you should set up the tests so that that warning is expected, handled, and the test fails if the warning is not raised (e.g., using [pytest.warns()](https://docs.pytest.org/en/3.0.5/recwarn.html)).

  Regarding other warnings (e.g., RuntimeWarning, DeprecationWarning), there are lots of options for warning handling in testing tools (I'm mostly familiar with those in pytest) but I think that how one uses them is a matter of developer preference (e.g., you could turn them all off using a configuration file, but perhaps you want to know when DeprecationWarnings arrive).  Ultimately this means that it is important for the developer to communicate to someone who might use the tests about what they should see.

Happy to discuss these and other points. Please let me know how I can assist as you respond and clarify reviewer comments.
