### 2020-10-01 0b25996
Addition of diagrams in documentation
- 3 diagrams for general layout of pvps, simulation and sizing process added to the scope section of the documentation.

- 1 diagram of the general layout of pvps added to readme.md
