### 2020-10-01 c23c963
Docstrings addition to pvgeneration.py and sizing.py
- Clarification on overwritting of surface tilt and azimuth by orientation_strategy in docstring

- Clarification of sizing process in sizing_minimize_npv()

- Cleaning of useless attributes in subset_respecting_llp_*()

- Addition of 2 paths for future development as comments in sizing.py
