### 2020-10-01 a0b50b6
Documentation improvements following SamuelDuchesne review
- addition of diagrams in scope of package_overview, and in .ipynb examples in order to clarify what a simulation and a sizing do.

- modifications in paper.md
