### 2020-10-02 #2637 @robinroche
> What does this "Getting started" page should contain more exactly in your opinion? Help on how to run the python and jupyter notebook examples? More general explanations around what is a PV pumping system (so far, I assume that the users have a minimum knowledge about PV pumping systems)? A more abstract layout of the steps of simulation than what is proposed in the examples? Thanks in advance for providing further details

@tylunel Yes, I was thinking of a high level explanation of how to use the code, and pointing the user to the basic example at the end of the installation.

> More information on the automated test script should be provided to the user. What can the user expect to see when running it? I got 26 warnings (see below)

@tylunel I agree with @kbarnhart . I did not notice anything worrying in the warnings, but I think that the user should know that getting these warnings may be normal. Otherwise one may wonder if the code will be fully operational or not.
