### 2020-10-02 ee1b9eb
Changes in instructions for installation. To be checked again after package is effectively added on pip.
