### 2020-10-07 0f07eba
Documentation improvement for the consideration or not of friction head
- renaming of variable 'iteration' into 'friction' in pvpumpsystem.py

- renaming of 'operating_point_noiteration()' into 'operating_point()' in pvpumpsystem.py

- adaptation of the rest of the code to these new names

- Docstring addition to clarify use of parameter 'friction'

- Jupyter notebook expanded to clarify use of 'friction'.
