### 2020-10-08 de88cf2
Handling of warnings in the tests
- All UserWarning are now captured and checked in the tests

- OptimizeWarning and RuntimeWarning ignored where justified

- minor formatting fixes
