+++ pvpumpingsystem/test/test_pump.py
@@ -12,9 +12,6 @@ import inspect
 import pvpumpingsystem.pump as pp
 from pvpumpingsystem import errors
 
-# ignore all OptimizeWarning in the module
-pytestmark = pytest.mark.filterwarnings(
-    "ignore::scipy.optimize.OptimizeWarning")
 
 test_dir = os.path.dirname(
     os.path.abspath(inspect.getfile(inspect.currentframe())))
@@ -25,28 +22,48 @@ def pumpset():
     pump_testfile = os.path.join(test_dir,
                                  '../data/pump_files/SCB_10_150_120_BL.txt')
     return pp.Pump(path=pump_testfile,
-                   idname='SCB_10',
-                   modeling_method='arab',
-                   motor_electrical_architecture='permanent_magnet')
+                   modeling_method='arab')
 
 
-def test_init(pumpset):
-    assert type(pumpset.specs) is pd.DataFrame
-    assert type(pumpset.coeffs['coeffs_f1']) is np.ndarray
+def test_init_pumpset_overwritting():
+    pump_testfile = os.path.join(test_dir,
+                                 '../data/pump_files/SCB_10_150_120_BL.txt')
+    with pytest.warns(UserWarning) as record:
+        pump1 = pp.Pump(path=pump_testfile,
+                        idname='SCB_10',
+                        price=1050,
+                        modeling_method='arab',
+                        motor_electrical_architecture='permanent_magnet')
+
+    # check that 2 warnings were raised
+    assert len(record) == 3
+    # check that the messages match
+    assert "price attribute overwritten." in record[0].message.args[0]
+    assert "idname attribute overwritten." in record[1].message.args[0]
+    assert "architecture attribute overwritten." in record[2].message.args[0]
+
+    # check that object is correctly built
+    assert type(pump1.specs) is pd.DataFrame
+    assert type(pump1.coeffs['coeffs_f1']) is np.ndarray
+
 
 
 def test_limited_pump_data_1():
     """Tests 'theoretical_basic' model.
-
     This pump data has only one data point"""
 
     pump_testfile = os.path.join(
             test_dir, '../data/pump_files/min_specs.txt')
     # The pump is modeled with extremely basic model, considering only f2
-    pumpset = pp.Pump(path=pump_testfile,
-                      modeling_method='theoretical',
-                      motor_electrical_architecture='permanent_magnet')
+    with pytest.warns(UserWarning) as record:
+        pumpset = pp.Pump(path=pump_testfile,
+                          modeling_method='theoretical',
+                          motor_electrical_architecture='permanent_magnet')
+    # check that the warnings are the expected ones
+    assert "architecture attribute overwritten." in record[0].message.args[0]
+    assert "recomputed with constant efficiency" in record[1].message.args[0]
 
+    # check functions can work normally then
     functQ, intervals = pumpset.functQforPH()
     res = functQ(540, 50)['Q']
     res_expected = 2.1
@@ -69,9 +86,15 @@ def test_limited_pump_data_2():
             test_dir, '../data/pump_files/rosen_SC33-158-D380-9200.txt')
     # The initialization recalculates the current used based on a constant
     # efficiency
-    pumpset = pp.Pump(path=pump_testfile,
-                      modeling_method='theoretical',
-                      motor_electrical_architecture='permanent_magnet')
+    with pytest.warns(UserWarning) as record:
+        pumpset = pp.Pump(path=pump_testfile,
+                          modeling_method='theoretical',
+                          motor_electrical_architecture='permanent_magnet')
+
+    # check that the warnings are the ones expected
+    assert "recomputed with constant efficiency" in record[0].message.args[0]
[…]
