### 2020-10-09 d12edb9
New section 'Getting started' added to documentation
- section expanded from 'examples' ('examples' deleted)

- adaptation in other files
