+++ docs/source/generated/pvpumpingsystem.pvpumpsystem.operating_point.rst
@@ -1,0 +1,6 @@
+pvpumpingsystem.pvpumpsystem.operating\_point
+=============================================
+
+.. currentmodule:: pvpumpingsystem.pvpumpsystem
+
+.. autofunction:: operating_point
