### 2020-10-09 01a3d87
Documentation expanded
- Explanations added on source of data (PV module, pump, weather files)

- paper.md expanded with comparison between pvpumpingsystem and PVsyst and SISIFO

- version set at 0.9
