+++ paper/paper.md
@@ -3,7 +3,7 @@ title: 'pvpumpingsystem: a python package for modeling and sizing photovoltaic w
 tags:
   - Python
   - sizing
-  - modeling 
+  - modeling
   - water pumping
   - photovoltaics
   - solar energy
@@ -25,50 +25,55 @@ bibliography: paper.bib
 
 # Summary
 
-According to the World Health Organisation, one tenth of the world’s population still lacks access to 
-basic water supply. One of the reasons for this is the remoteness of these populations from modern 
-water collection and distribution technologies, often coupled with an unfavourable socio-economic 
-situation. Photovoltaic (PV) pumping technology makes it possible to respond both to this problem 
-and to the criteria of sustainable development. However, these pumping systems must be carefully 
+According to the World Health Organisation, one tenth of the world population still lacks access to
+basic water supply. One of the reasons for this is the remoteness of these populations from modern
+water collection and distribution technologies, often coupled with an unfavourable socio-economic
+situation. Photovoltaic (PV) pumping technology makes it possible to respond both to this problem
+and to the criteria of sustainable development. However, these pumping systems must be carefully
 modeled and sized in order to make the water supply cost efficient and reliable.
 
-Pvpumpingsystem was conceived in order to tackle this issue. It is an open source package 
-providing various tools aimed at facilitating the modeling and the sizing of photovoltaic 
-powered water pumping systems. Even though the package is originally targeted at researchers 
+Pvpumpingsystem was conceived in order to tackle this issue. It is an open source package
+providing various tools aimed at facilitating the modeling and the sizing of photovoltaic
+powered water pumping systems. Even though the package is originally targeted at researchers
 and engineers, three practical examples are provided in order to help anyone to use pvpumpingsystem.
 
-Python is the programming language used in the software, and the code is structured within an 
-object-oriented approach. Continuous integration services allow checking for lint in the code 
-and to automatize the tests. Each class and function are documented with reference to the 
+Python is the programming language used in the software, and the code is structured within an
+object-oriented approach. Continuous integration services allow checking for lint in the code
+and to automatize the tests. Each class and function are documented with reference to the
 literature when applicable. Pvpumpingsystem is released under a GPL-v3 license.
 
-Pvpumpingsystem relies on already existing packages for photovoltaic and fluid mechanics modeling, 
-namely “pvlib-python” [@pvlib-python] and “fluids” [@fluids]. pvpumpingsystem’s originality lies 
-in the implementation of various motor-pump models for finite power sources and in the coupling 
-of the distinct component models. In order to increase the understandability of the code, 
-each physical component of the PV pumping system corresponds to a class, like for example 
-the classes Pump(), MPPT(), PipeNetwork(), Reservoir() and PVGeneration(). The previous objects 
-are then gathered in the class PVPumpSystem() which allows running a comprehensive modeling of 
-the pumping system. 
+Pvpumpingsystem relies on already existing packages for photovoltaic and fluid mechanics modeling,
+namely “pvlib-python” [@pvlib-python] and “fluids” [@fluids]. pvpumpingsystem originality lies
+in the implementation of various motor-pump models for finite power sources and in the coupling
+of the distinct component models. In order to increase the understandability of the code,
+each physical component of the PV pumping system corresponds to a class, like for example
+the classes Pump(), MPPT(), PipeNetwork(), Reservoir() and PVGeneration(). The previous objects
+are then gathered in the class PVPumpSystem() which allows running a comprehensive modeling of
+the pumping system.
 
-The main inputs to the simulation are hourly weather file, water source characteristics, expected water 
-consumption profile, and specifications of photovoltaic array, motor-pump and water reservoir. 
-Typical outputs are hourly flow rates, unused electric power, efficiency of components, life 
[…]
