### 2020-10-12 af39351
installation instructions modified
- instruction on installation from pip fixed
