### 2020-10-12 a7dfa5e
installation instructions modified
- instruction in README.md changed
