### 2020-10-12 #2637 @tylunel
The changes have been made as proposed in the reviews. I hope that it corresponds well to your expectations @robinroche @samuelduchesne.

There is one point that was maybe not addressed completely as expected, it is for the clear definition of the 2 use cases (penultimate item of @samuelduchesne). On that, I did not want to insist too much on the sizing case, as it is not the core of the package. The modeling tools are the core, and the sizing tool proposed is more a wrapper of the previous tools, and should be more seen as an example of how the code can be used to adress complex problems. I tried to clarify it in the docs as well.

For the rest of the modifications, there were all useful I think, thanks!
