### 2020-10-13 #2637 @samuelduchesne
Thank you @tylunel for addressing these comments. I have nothing to add 😀
