### 2020-10-13 #2637 @kbarnhart
@robinroche and @samuelduchesne thank you for your thoughtful comments and timely review. @tylunel thank you for responding to the reviewer comments.

Now that the reviewers have completed the checklist and recommended publication the next step is that I will do a final proofreading step on the publication (I may make a PR or just provide a list). This will happen in the next few days. Once that is complete I will ask @tylunel to complete a few steps to create the final archive and will then pass the submission on to the JOSS Editors in Chief for final processing.
