### 2020-10-15 #2637 @kbarnhart
@tylunel it looks like there is an issue with the paper.md file that makes it not compile. Would you address it and then ping me so I can proceed.

If you make changes on a test branch, know that you can use whedon to compile off a custom branch:
