### 2020-10-21 25ca3fe
Format change (UTF-8) for paper.md
