### 2020-10-21 #2637 @tylunel
@kbarnhart Sorry for that, I forgot to check the encoding in UTF-8 after the last changes in paper.md.
It should be okay now!
