+++ paper/paper.bib
@@ -10,10 +10,9 @@
 
 @misc{fluids,
     author = {Caleb Bell},
-    title = {fluids: Fluid dynamics component of Chemical Engineering Design Library (ChEDL)},
+    title = {{fluids: Fluid dynamics component of Chemical Engineering Design Library (ChEDL)}},
     year = {2020},
     publisher = {GitHub},
     journal = {GitHub repository},
     url = {https://github.com/CalebBell/fluids}
 }
-
