### 2020-10-22 491533d
Minor repetition fix
repetition of 'in' removed in paper.md
