### 2020-10-22 #2637 @kbarnhart
@tylunel thanks for merging those copyedits. Congratulations on making to the end of the JOSS review process.

At this point could you:
