### 2020-10-26 #2637 @tylunel
The release version is v1.0.0 (https://github.com/tylunel/pvpumpingsystem/releases/tag/v1.0.0) and the DOI of the archive version is 10.5281/zenodo.4134280 (https://zenodo.org/record/4134280).
