### 2020-10-26 #2637 @kbarnhart
@tylunel this is now in the hands of the JOSS handling Editor in Chief who manages final publication.
